api/requirements.txt
django
django-rest-framework
pillow
django-oauth-toolkit
dj-rest-auth
djangorestframework-simplejwt
